pyproject.toml
[tool.poetry]
name = "django-proj-task"
version = "0.1.0"
description = "Take home task"
authors = ["Tomislav Nazifović <tomislav.nazifovic@gmail.com>"]
readme = "README.md"
#packages = [{include = "django_proj_task", from="home_budget"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.2.4"
djangorestframework = "^3.16.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
mypy = "^1.16.1"
ruff = "^0.12.2"
pytest-django = "^4.11.1"
pytest-mypy = "^1.0.1"
pytest-ruff = "^0.5"
pytest-cov = "^6.2.1"
django-stubs = "^5.2.1"
djangorestframework-stubs = "^3.16.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = "-rfxX --mypy --ruff --verbose --cov-report html --cov-report term-missing --cov=budget_app --cov=home_budget"
testpaths = [
    "tests",
    "budget_app",
    "home_budget",
]
markers = [
    "ruff",
    "mypy",
]
DJANGO_SETTINGS_MODULE = "home_budget.settings"
rootdir = "home_budget"

[tool.coverage.run]
omit = ["budget_app/migrations/*"]

[tool.ruff]
exclude = [".git", ".pytest_cache", ".mypy_cache", ".vscode"]
line-length = 100
target-version = "py311"
output-format = "full"

[tool.ruff.lint]
select = ["E", "W", "F", "C", "B", "I", "N"]

[tool.ruff.lint.isort]
combine-as-imports = true
